move internal/proxynode to internal/proxy
