+++ internal/proxy/channels_mgr_test.go
