+++ internal/proxy/conf_adapter.go
