+++ internal/proxy/conf_adapter_mgr.go
