+++ internal/proxy/conf_adapter_mgr_test.go
