+++ internal/proxy/conf_adapter_test.go
