+++ internal/proxy/meta_cache_test.go
