+++ internal/proxy/segment.go
